Refactor (#1)
* add gpu code

* carvanha codegen

* average softmax and more stats

* massive update

* forward + softmax works

* fixed seed and cpu fwd activation matches gpu

* fix: push 3layer code

* local truncation emulation in cleartext and pvgg16

* reluextend in cpu code

* new truncation design, somewhat tested

* wip commit

* wip 2

* reluextend hotfix, might still be wrong?

* relu extend probably works now

* minor change

* llama improved and carvahna

* label subtraction in gpu

* two round sign extension

* save before full run

* these bugs make me wonder my sanity

* update main

* fix maxpool timing bug

* deterministic dealer

* hot fix

* save progress

* some changes

* delete perf files

* benchmarking commit

* rt microbenchmark

* save code

* minor fix

+++ backend/minillama/utils.h
@@ -51,7 +51,7 @@ void MatFinalize4(int bw, int s1, int s2, int s3, int s4, GroupElement *input);
 
 void Conv2DReshapeFilter(int FH, int FW, int CI, int CO, GroupElement* filter, eigenMatrix &reshapedFilter);
 
-void Conv2DReshapeInput(int N, int H, int W, int CI, int FH, int FW, int zPadHLeft, int zPadHRight, int zPadWLeft, int zPadWRight, int strideH, int strideW, int RRows, int RCols, GroupElement *inputArr, GroupElement *outputArr);
+void Conv2DReshapeInput(size_t N, size_t H, size_t W, size_t CI, size_t FH, size_t FW, size_t zPadHLeft, size_t zPadHRight, size_t zPadWLeft, size_t zPadWRight, size_t strideH, size_t strideW, size_t RRows, size_t RCols, GroupElement *inputArr, GroupElement *outputArr);
 
 void Conv2DReshapeOutput(int N, int finalH, int finalW, int CO, GroupElement *inputArr, GroupElement *outputArr);
